workflow/requirements.txt
bcrypt==4.1.2
PyJWT==2.8.0
boto3==1.34.80
pypdf==4.1.0
